refactor manager user
